third_party/blink/web_tests/fast/history/change-viewport-height-in-onpopstate-expected.txt

CONSOLE MESSAGE: scrollY is: 0
